Backend & Frontend work
I have redirected the server to work on `server.py` file in order to handle all the URL links and all.

templates folder has been removed to due to the change of directories to the main one.

New page and url added for upcomining events with styling sheets. As well as adding new images to Static/Images directory.